fix(devtools): restore allowlist-based test layout invariant
Recent changes made check-test-layout fail the repo due to existing multi-binary integration test suites.

- Reintroduce test-layout allowlist support (and a --allowlist flag) so existing violators remain allowed while new ones are blocked
- Refresh test-layout-allowlist.txt to match the current noncompliant crates
- Keep nova-dap suite consolidated (module registration) and apply a few small clippy-driven cleanups

+++ crates/nova-memory/src/budget.rs
@@ -291,7 +291,7 @@ fn process_rlimit_as_bytes() -> Option<u64> {
         return None;
     }
 
-    interpret_rlimit_as_bytes(limit.rlim_cur as u64, libc::RLIM_INFINITY as u64)
+    interpret_rlimit_as_bytes(limit.rlim_cur, libc::RLIM_INFINITY)
 }
 
 #[cfg(not(unix))]
